+++ pypeman/tests/test_app/settings_disabled.py
@@ -1,0 +1,3 @@
+from pypeman.default_settings import *  # noqa F403
+
+DISABLED_PLUGINS = ["pypeman.plugins.graph.GraphPlugin"]