+++ pypeman/tests/tests_bin.py
@@ -1,8 +1,10 @@
 # pypeman -- A command-line argument parser for Python
 # Copyright (C) 2015-2016 by MHComm. See LICENSE for details
 
+import json
 import os
 import sys
+import tempfile
 import unittest
 import subprocess
 import logging
@@ -30,32 +32,30 @@ class BinPypemanTestCase(unittest.TestCase):
             if os.path.exists(fname):
                 os.unlink(fname)
 
-    def run_pypeman(self, cmd, cwd=None):
+    def run_pypeman(self, cmd, cwd=None, expected_code=0):
         """
         Runs a command, gathers output and captures exit code
-        :return: return_code and data (bytestring of stdout / stderr)
+        :return: return_code and data (bytestring of stdout)
         """
         out_fname = mktempfname()
-        self.tempfiles.append(out_fname)
+        err_fname = mktempfname()
+        self.tempfiles.extend((out_fname, err_fname))
 
         if cwd is None:
             cwd = os.getcwd()
 
-        # TODO: why not reading stdout / stderr directoy into a var ??
-        # TODO: if not reason can be found change code to use no temp file
-        # TODO: perhaps some issue with unicode / python 3 ???
-        with open(out_fname, 'wb') as fout:
-            proc = subprocess.Popen(cmd, stdout=fout, stderr=fout, cwd=cwd)
+        with open(out_fname, 'wb') as fout, open(err_fname, 'wb') as ferr:
+            proc = subprocess.Popen(cmd, stdout=fout, stderr=ferr, cwd=cwd)
             proc.wait()
         ret_code = proc.returncode
 
-        data = None
-        if ret_code:
-            with open(out_fname, 'rb') as fin:
-                data = fin.read()
+        with open(out_fname, 'rb') as fin:
+            data = fin.read()
+        with open(err_fname, 'rb') as fin:
+            err_data = fin.read()
 
-        self.assertEqual(ret_code, 0, "exit code %d when calling %r in %s: %s" %
-                         (ret_code, cmd, cwd, data))
+        self.assertEqual(ret_code, expected_code, "exit code %d when calling %r in %s: %s / %s" %
+                         (ret_code, cmd, cwd, data, err_data))
 
         return ret_code, data
 
@@ -68,15 +68,50 @@ class BinPypemanTestCase(unittest.TestCase):
         self.run_pypeman(cmd, cwd=CWD)
 
     def test_02_can_call_graph(self):
-        """ subcommand graph is working """
+        """ subcommand graph is working (all formats) """
 
-        cmd = self.cmd + ['graph']
-        self.run_pypeman(cmd, cwd=CWD)
+        self.run_pypeman(self.cmd + ['graph'], cwd=CWD)
+        self.run_pypeman(self.cmd + ['graph', '--format', 'dot'], cwd=CWD)
 
-    def test_03_can_call_test(self):
-        """ subcommand test is working """
+        _, data = self.run_pypeman(self.cmd + ['graph', '--format', 'json'], cwd=CWD)
+        graph = json.loads(data)
+        self.assertEqual(graph['version'], 1)
+        self.assertIn('channels', graph)
 
-        cmd = self.cmd + ['test']
-        self.run_pypeman(cmd, cwd=os.path.join(os.path.dirname(CWD), 'test_app_testing'))
+    def test_03_can_call_listplugins_and_printsettings(self):
+        """ subcommands listplugins and printsettings are working """
 
-# test_suite =  BinPypemanTestCase
+        _, data = self.run_pypeman(self.cmd + ['listplugins'], cwd=CWD)
+        self.assertIn(b'GraphPlugin', data)
+
+        _, data = self.run_pypeman(self.cmd + ['printsettings'], cwd=CWD)
+        self.assertIn(b'PROJECT_MODULE', data)
+
+    def test_04_can_deactivate_plugins(self):
+        """ DISABLED_PLUGINS drops the plugin's command and listplugins reports it """
+
+        os.environ['PYPEMAN_SETTINGS_MODULE'] = 'settings_disabled'
+        try:
+            _, data = self.run_pypeman(self.cmd + ['listplugins'], cwd=CWD)
+            # not active any more (active entries read "<module> GraphPlugin:")
+            self.assertNotIn(b'GraphPlugin:', data)
+            self.assertIn(b'Deactivated by settings.DISABLED_PLUGINS:', data)
+            self.assertIn(b'pypeman.plugins.graph.GraphPlugin', data)
+
+            # its command is gone: argparse rejects it (exit code 2)
+            self.run_pypeman(self.cmd + ['graph'], cwd=CWD, expected_code=2)
+        finally:
+            del os.environ['PYPEMAN_SETTINGS_MODULE']
+
+    def test_05_can_call_startproject(self):
+        """ bin/pypeman-startproject creates the project files """
+
+        script = os.path.join(os.path.dirname(__file__), '..', '..', 'bin', 'pypeman-startproject')
+        with tempfile.TemporaryDirectory() as tmpdir:
+            # run from a NON-project cwd: no settings module around
+            self.run_pypeman([sys.executable, script, 'demo_pjt'], cwd=tmpdir)
+
+            pjt_dir = os.path.join(tmpdir, 'demo_pjt')
+            for fname in ('project.py', 'settings.py', 'dist_settings.py', 'tests.py'):
+                self.assertTrue(os.path.exists(os.path.join(pjt_dir, fname)),
+                                "missing generated file %s" % fname)